Fixed bug where runes' data was not getting save to properly in the LivingEntitiesData.cs.
id was not getting incremented.
the for loop wasn't set to start at start ids 3 change to start ids 2.
the for loop wasn't set to end at the right place ended and start ids 2 change to start ids 3.
All Items was getting the wrong index id was being used as the indexer changed to i from the the for loop.