Created a struct that hold 2 ints.
Inventory now has an array of the ranges struct.

Inventory now loops though the inventory state enum for the different types of items to be randomly generated.

Now called roll on Roller that passes in the InventoryState enum to tell the Roller what kind of item to make.

AllItems in the InventoryState Enum was changed to be the last element instead of the first.

DO so the inventory can loop though all the item types and stop at the right place.

Changed all calls to set Inventory that were passing in 0 to InvetoryState.ALlItems.

All rollers under Roller.roller are now private.

Roller.roller now has function calls for the creation of items from there crafting station.