Fixed issue with the forge ui not passing in the smithing level of the player.
Started making all fields on weapon holder private with get functions.

Done so that the values of the weapon can not be changed after its created.

Started a function that will set the starting values for the weapon. IF the values are already not null then the function returns without doing anything.